% Auto-generated: do not edit by hand
\name{carousel}

\alias{carousel}

\title{Carousel component}

\description{
ECarousel is an carousel component. It takes some setting properties You put how many blocks of divs you want inside it, and this div turn into slides
}

\usage{
carousel(children=NULL, dots=NULL, arrows=NULL, infinite=NULL,
autoplay=NULL, speed=NULL, slidesToShow=NULL,
slidesToScroll=NULL, centerMode=NULL, centerPadding=NULL,
swipeToSlide=NULL, variableWidth=NULL, responsive=NULL,
vertical=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Your carousel is vertical}

\item{dots}{Logical. Dots under carousel}

\item{arrows}{Logical. Arrows to control carousel}

\item{infinite}{Logical. If the carousel content will repeat in a loop}

\item{autoplay}{Logical. If the carousel will start to play automatically}

\item{speed}{Numeric. Speed of autoplay}

\item{slidesToShow}{Numeric. How many slides you want to show}

\item{slidesToScroll}{Numeric. How many slides will scroll when you swipe the carousel}

\item{centerMode}{Logical. To centralize the carousel}

\item{centerPadding}{Character. Padding on the sides}

\item{swipeToSlide}{Logical. If you can slide to scroll the slides}

\item{variableWidth}{Logical. The slides width varies according to the screen size}

\item{responsive}{Unnamed list. Settings of breakpoints}

\item{vertical}{Logical. If your carousel is vertical}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

