% Auto-generated: do not edit by hand
\name{gridLayoutComponent}

\alias{gridLayoutComponent}

\title{GridLayoutComponent component}

\description{
GridLayout is an GridLayout component. It takes some setting properties You put how many blocks of divs you want inside it, and this div turn into slides
}

\usage{
gridLayoutComponent(children=NULL, style=NULL, className=NULL, id=NULL,
layout=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Your GridLayout is vertical}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{id}{Character. Id of the element}

\item{layout}{Unnamed list. Initial layout}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

