% Auto-generated: do not edit by hand
\name{sideBarToggle}

\alias{sideBarToggle}

\title{SideBarToggle component}

\description{

}

\usage{
sideBarToggle(n_clicks=NULL, id=NULL)
}

\arguments{
\item{n_clicks}{Numeric. How many slides you want to show}

\item{id}{Character. Padding on the sides}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

