% Auto-generated: do not edit by hand
\name{themeToggle}

\alias{themeToggle}

\title{ThemeToggle component}

\description{

}

\usage{
themeToggle(bg_color_dark=NULL, icon_color_dark=NULL,
bg_color_light=NULL, icon_color_light=NULL,
tooltip_text=NULL, style=NULL, className=NULL, id=NULL)
}

\arguments{
\item{bg_color_dark}{Character. Background color of toggle switch when dark theme}

\item{icon_color_dark}{Character. Sun icon color of toggle switch when dark theme}

\item{bg_color_light}{Character. Background color of toggle switch when light theme}

\item{icon_color_light}{Character. Moon icon color of toggle switch when dark theme}

\item{tooltip_text}{Character. Text that will appear in tooltip (only desktop)}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{id}{Character. Id of the element}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

