% Auto-generated: do not edit by hand
\name{carousel}

\alias{carousel}

\title{Carousel component}

\description{

}

\usage{
carousel(children=NULL, dots=NULL, arrows=NULL, infinite=NULL,
autoplay=NULL, speed=NULL, slidesToShow=NULL,
slidesToScroll=NULL, centerMode=NULL, centerPadding=NULL,
swipeToSlide=NULL, variableWidth=NULL, responsive=NULL,
vertical=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{dots}{Logical. }

\item{arrows}{Logical. }

\item{infinite}{Logical. }

\item{autoplay}{Logical. }

\item{speed}{Numeric. }

\item{slidesToShow}{Numeric. }

\item{slidesToScroll}{Numeric. }

\item{centerMode}{Logical. }

\item{centerPadding}{Character. }

\item{swipeToSlide}{Logical. }

\item{variableWidth}{Logical. }

\item{responsive}{Unnamed list. }

\item{vertical}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

