% Auto-generated: do not edit by hand
\name{carousel}

\alias{carousel}

\title{Carousel component}

\description{
ECarousel is an carousel component. It takes some setting properties You put how many blocks of divs you want inside it, and this div turn into slides
}

\usage{
carousel(children=NULL, dots=NULL, arrows=NULL, infinite=NULL,
autoplay=NULL, speed=NULL, slides_to_show=NULL,
slides_to_scroll=NULL, center_mode=NULL,
center_padding=NULL, swipe_to_slide=NULL,
variable_width=NULL, responsive=NULL, vertical=NULL,
style=NULL, className=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Your carousel is vertical}

\item{dots}{Logical. Dots under carousel}

\item{arrows}{Logical. Arrows to control carousel}

\item{infinite}{Logical. If the carousel content will repeat in a loop}

\item{autoplay}{Logical. If the carousel will start to play automatically}

\item{speed}{Numeric. Speed of autoplay}

\item{slides_to_show}{Numeric. How many slides you want to show}

\item{slides_to_scroll}{Numeric. How many slides will scroll when you swipe the carousel}

\item{center_mode}{Logical. To centralize the carousel}

\item{center_padding}{Character. Padding on the sides}

\item{swipe_to_slide}{Logical. If you can slide to scroll the slides}

\item{variable_width}{Logical. The slides width varies according to the screen size}

\item{responsive}{Unnamed list. Settings of breakpoints}

\item{vertical}{Logical. If your carousel is vertical}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{id}{Character. Id of the element}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

