% Auto-generated: do not edit by hand
\name{sideBar}

\alias{sideBar}

\title{SideBar component}

\description{

}

\usage{
sideBar(children=NULL, id=NULL, style=NULL, className=NULL,
bg_color=NULL, text_color=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Your carousel is vertical}

\item{id}{Character. Id of the element}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{bg_color}{Character. Color of sidebar background}

\item{text_color}{Character. Color of sidebar text}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

